// src/frontend/src/src/backend.d.ts
export type Time = bigint;
// ...
export interface Employee {
    id: bigint;
    status: string;
    salary: number;
    hireDate: Time;
    name: string;
    role: string;
    email: string;
    department: string;
}